Add files via upload
Using rasa, I'm building Moby's brain. I used a wikipedia api for this however, there seems to be a problem with where the files are saved, as when i try to train and test the bot, it doesn't respond to the changed i've made but only goes along with the default settings.